+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2021-03-14
+### Changed
+* Allowing clearing only specific metrics from meter cache.
+* Minor optimization in transactions helper.
+
 ## [1.4.0] - 2021-03-01
 ### Added
 * MeterCache to avoid Micrometer's MeterRegistry's inefficiencies.